requirements.txt
numpy
reportlab
Pillow
playwright
